Move script into legacy-versions.ts
